fix process_mask

+++ utils/segment/general.py
@@ -44,7 +44,7 @@ def process_mask_upsample(proto_out, out_masks, bboxes, shape):
     """
 
     c, mh, mw = proto_out.shape  # CHW
-    masks = (out_masks.tanh() @ proto_out.view(c, -1)).sigmoid().view(-1, mh, mw)
+    masks = (out_masks.tanh() @ proto_out.float().view(c, -1)).sigmoid().view(-1, mh, mw)
     masks = F.interpolate(masks[None], shape, mode='bilinear', align_corners=False)[0]  # CHW
     masks = crop(masks.permute(1, 2, 0).contiguous(), bboxes)  # HWC
     return masks.gt_(0.5)
@@ -63,7 +63,7 @@ def process_mask(proto_out, out_masks, bboxes, shape, upsample=False):
 
     c, mh, mw = proto_out.shape  # CHW
     ih, iw = shape
-    masks = (out_masks.tanh() @ proto_out.view(c, -1)).sigmoid().view(-1, mh, mw)  # CHW
+    masks = (out_masks.tanh() @ proto_out.float().view(c, -1)).sigmoid().view(-1, mh, mw)  # CHW
 
     downsampled_bboxes = bboxes.clone()
     downsampled_bboxes[:, 0] *= mw / iw
